requirements.txt
ctrader-open-api==0.9.2
protobuf==3.20.1
Twisted==24.3.0
python-dotenv
flask
